Multi-tenency. v.9

+++ src/main/java/org/servantscode/person/db/FamilyDB.java
@@ -4,6 +4,7 @@ import org.servantscode.commons.db.DBAccess;
 import org.servantscode.commons.db.ReportStreamingOutput;
 import org.servantscode.commons.search.QueryBuilder;
 import org.servantscode.commons.search.SearchParser;
+import org.servantscode.commons.security.OrganizationContext;
 import org.servantscode.person.Address;
 import org.servantscode.person.Family;
 
@@ -25,7 +26,7 @@ public class FamilyDB extends DBAccess {
     }
 
     public int getCount(String search, boolean includeInactive) {
-        QueryBuilder query = count().from("families").search(searchParser.parse(search));
+        QueryBuilder query = count().from("families").search(searchParser.parse(search)).inOrg();
         if(!includeInactive)
             query.where("inactive=false");
 
@@ -40,7 +41,7 @@ public class FamilyDB extends DBAccess {
     }
 
     public StreamingOutput getReportReader(String search, boolean includeInactive, final List<String> fields) {
-        final QueryBuilder query = selectAll().from("families").search(searchParser.parse(search));
+        final QueryBuilder query = selectAll().from("families").search(searchParser.parse(search)).inOrg();
         if(!includeInactive)
             query.where("inactive=false");
 
@@ -60,7 +61,7 @@ public class FamilyDB extends DBAccess {
     }
 
     public List<Family> getFamilies(String search, String sortField, int start, int count, boolean includeInactive) {
-        QueryBuilder query = selectAll().from("families").search(searchParser.parse(search));
+        QueryBuilder query = selectAll().from("families").search(searchParser.parse(search)).inOrg();
         if(!includeInactive)
             query.where("inactive=false");
         query.sort(sortField).limit(count).offset(start);
@@ -76,7 +77,7 @@ public class FamilyDB extends DBAccess {
     }
 
     public Family getFamily(int id) {
-        QueryBuilder query = selectAll().from("families").where("id=?", id);
+        QueryBuilder query = selectAll().from("families").where("id=?", id).inOrg();
 
         try ( Connection conn = getConnection();
               PreparedStatement stmt = query.prepareStatement(conn)
@@ -91,7 +92,7 @@ public class FamilyDB extends DBAccess {
 
     public void create(Family family) {
         try (Connection conn = getConnection();
-             PreparedStatement stmt = conn.prepareStatement("INSERT INTO families(surname, home_phone, envelope_number, addr_street1, addr_street2, addr_city, addr_state, addr_zip) values (?, ?, ?, ?, ?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS)
+             PreparedStatement stmt = conn.prepareStatement("INSERT INTO families(surname, home_phone, envelope_number, addr_street1, addr_street2, addr_city, addr_state, addr_zip, org_id) values (?, ?, ?, ?, ?, ?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS)
         ){
             Address addr = family.getAddress();
 
@@ -103,6 +104,7 @@ public class FamilyDB extends DBAccess {
             stmt.setString(6, addr.getCity());
             stmt.setString(7, addr.getState());
             stmt.setInt(8, addr.getZip());
+            stmt.setInt(9, OrganizationContext.orgId());
 
             if(stmt.executeUpdate() == 0) {
                 throw new RuntimeException("Could not create family: " + family.getSurname());
@@ -119,7 +121,7 @@ public class FamilyDB extends DBAccess {
 
     public void update(Family family) {
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("UPDATE families SET surname=?, home_phone=?, envelope_number=?, addr_street1=?, addr_street2=?, addr_city=?, addr_state=?, addr_zip=?, inactive=? WHERE id=?")
+              PreparedStatement stmt = conn.prepareStatement("UPDATE families SET surname=?, home_phone=?, envelope_number=?, addr_street1=?, addr_street2=?, addr_city=?, addr_state=?, addr_zip=?, inactive=?, org_id=? WHERE id=?")
             ){
 
             Address addr = family.getAddress();
@@ -133,7 +135,8 @@ public class FamilyDB extends DBAccess {
             stmt.setString(7, addr.getState());
             stmt.setInt(8, addr.getZip());
             stmt.setBoolean(9, family.isInactive());
-            stmt.setInt(10, family.getId());
+            stmt.setInt(10, OrganizationContext.orgId());
+            stmt.setInt(11, family.getId());
 
             if(stmt.executeUpdate() == 0)
                 throw new RuntimeException("Could not update family: " + family.getSurname());
@@ -145,10 +148,11 @@ public class FamilyDB extends DBAccess {
 
     public boolean delete(Family family) {
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("DELETE FROM families WHERE id=?")
+              PreparedStatement stmt = conn.prepareStatement("DELETE FROM families WHERE id=? AND org_id=?")
             ){
 
             stmt.setInt(1, family.getId());
+            stmt.setInt(2, OrganizationContext.orgId());
 
             return stmt.executeUpdate() != 0;
         } catch (SQLException e) {
@@ -158,11 +162,12 @@ public class FamilyDB extends DBAccess {
 
     public boolean deactivate(Family family) {
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("UPDATE families SET inactive=? WHERE id=?")
+              PreparedStatement stmt = conn.prepareStatement("UPDATE families SET inactive=? WHERE id=? AND org_id=?")
         ){
 
             stmt.setBoolean(1, true);
             stmt.setInt(2, family.getId());
+            stmt.setInt(3, OrganizationContext.orgId());
 
             return stmt.executeUpdate() != 0;
         } catch (SQLException e) {
@@ -172,10 +177,11 @@ public class FamilyDB extends DBAccess {
 
     public void attchPhoto(int id, String guid) {
         try ( Connection conn = getConnection();
-              PreparedStatement stmt = conn.prepareStatement("UPDATE families SET photo_guid=? WHERE id=?");
+              PreparedStatement stmt = conn.prepareStatement("UPDATE families SET photo_guid=? WHERE id=? AND org_id=?");
         ){
             stmt.setString(1, guid);
             stmt.setInt(2, id);
+            stmt.setInt(3, OrganizationContext.orgId());
 
             if(stmt.executeUpdate() == 0)
                 throw new NotFoundException("Could not attach photo to family: " + id);
